moved project.js and individualproject.js to view folder
