+++ blitz-build/src/views/projects/IndividualProject.js
@@ -3,13 +3,13 @@ import { withRouter } from "react-router-dom";
 import { axiosWithAuth } from "../../utils/auth/axiosWithAuth";
 
 //components
-import Weather from "../weather/Weather";
-import ProjectTaskCard from "./ProjectsTaskCard";
+import Weather from "../../components/weather/Weather";
+import ProjectTaskCard from "../../components/projects/ProjectsTaskCard";
 import Global from "../../styles/Global";
-import DeleteProject from "./DeleteProject";
-import EditProject from "./EditProject";
-import Documents from "../documents/Documents";
-import ProjectTasks from "../../views/tasks/ProjectTasks";
+import DeleteProject from "../../components/projects/DeleteProject";
+import EditProject from "../../components/projects/EditProject";
+import Documents from "../../components/documents/Documents";
+import ProjectTasks from "../tasks/ProjectTasks";
 
 //static
 import Project_icon from "../../styles/icons_project/project_icon.png";
